deploy refs/heads/master b6b53745b2624ed60eb9902f2f367e6c25a4e04a

+++ dist.sql
@@ -915,7 +915,7 @@ drop table if exists chii_likes;
 create table IF NOT EXISTS `chii_likes`
 (
     `type`       MEDIUMINT UNSIGNED  NOT NULL,
-    `related_id` mediumint           NOT NULL,
+    `related_id` int(10) unsigned    NOT NULL,
     `uid`        mediumint UNSIGNED  NOT NULL,
     `value`      mediumint unsigned  not null default 0,
     `ban`        tinyint(2) unsigned not null DEFAULT 0,
